feat: refactor codes to get ready for UI
Split main.py into core.py and console.py in preparation for introducing the UI in the future
